Testing

+++ Lab #4/Code_03_500923949_ChengHao_Kao_LAB4/Code_03_500923949_ChengHao_Kao_LAB4.ino
@@ -5,16 +5,16 @@ Servo myservo;
 #define echoPin 5
 #define trigPin 4
 
-long duration;
-int distance_cm;
+float duration;
+float distance_cm;
 int pos = 0;
 
 void setup()
 {
   Serial.begin(9600);
   Serial.setTimeout(1);
-  pinMode(trigPin, OUTPUT); 
-  pinMode(echoPin, INPUT);  
+  pinMode(trigPin, OUTPUT);
+  pinMode(echoPin, INPUT);
   myservo.attach(6);
   Serial.println("Radar Start");
   //myservo.write(180);
@@ -22,9 +22,9 @@ void setup()
 
 void loop()
 {
-  for (int i = 0; i < 180; i++) 
+  for (int i = 0; i < 180; i++)
   {
-    pos+=1;
+    pos += 1;
     myservo.write(pos);
     digitalWrite(trigPin, LOW);
     delayMicroseconds(2);
@@ -32,15 +32,15 @@ void loop()
     delayMicroseconds(10);
     digitalWrite(trigPin, LOW);
     duration = pulseIn(echoPin, HIGH);
-    distance_cm = duration * 0.034 / 2;
+    distance_cm = duration * 0.034 / 2.0;
     Serial.print(pos);
     Serial.print(",");
     Serial.println(distance_cm);
     delay(50);
-  } 
-  for (int i = 180; i > 0; i--) 
+  }
+  for (int i = 180; i > 0; i--)
   {
-    pos-=1;
+    pos -= 1;
     myservo.write(pos);
     digitalWrite(trigPin, LOW);
     delayMicroseconds(2);
